requirements.txt
Flask==3.0.0
Flask-WTF==1.2.2
WTForms==3.2.1
gunicorn==21.2.0
numpy==1.26.4
pandas==2.1.1
scikit-learn==1.3.1
tensorflow-cpu==2.14.0
joblib==1.3.2
h5py==3.10.0
